+++ packages/platform-server/init/BUILD.bazel
@@ -1,5 +1,3 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-load("@npm//:rollup/package_json.bzl", rollup = "bin")
 load("//tools:defaults.bzl", "ng_project", "tsec_test")
 
 package(default_visibility = ["//visibility:public"])
@@ -14,35 +12,7 @@ ng_project(
         exclude = ["src/bundled-domino.d.ts"],
     ),
     deps = [
-        ":bundled_domino_lib",
-    ],
-)
-
-rollup.rollup(
-    name = "bundled_domino",
-    srcs = [
-        "//:node_modules/@rollup/plugin-commonjs",
-        "//:node_modules/domino",
-    ],
-    outs = [
-        "src/bundled-domino.mjs",
-    ],
-    args = [
-        "--format=esm",
-        "--plugin=@rollup/plugin-commonjs",
-        "--input=node_modules/domino/lib/index.js",
-        "--sourcemap=false",
-        "--file=packages/platform-server/init/src/bundled-domino.mjs",
-    ],
-    progress_message = "Bundling domino",
-    silent_on_success = True,
-)
-
-js_library(
-    name = "bundled_domino_lib",
-    srcs = [
-        "src/bundled-domino.d.ts",
-        ":bundled_domino",
+        "//packages/platform-server/third_party/domino:bundled_domino_lib",
     ],
 )
