--- packages/platform-server/init/src/bundled-domino.d.ts
@@ -1,11 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.dev/license
- */
-
-import domino from 'domino';
-
-export default domino;
