+++ packages/platform-server/init/test/shims_spec.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.dev/license
  */
-import domino from '../src/bundled-domino';
+import domino from '../../third_party/domino/bundled-domino';
 import {applyShims} from '../src/shims';
 
 describe('applyShims()', () => {
