+++ packages/platform-server/test/BUILD.bazel
@@ -43,7 +43,7 @@ ts_project(
         "//packages/platform-browser",
         "//packages/platform-browser/animations",
         "//packages/platform-server",
-        "//packages/platform-server:bundled_domino_lib",
+        "//packages/platform-server/third_party/domino:bundled_domino_lib",
         "//packages/private/testing",
         "//packages/router",
     ],
