+++ packages/private/testing/BUILD.bazel
@@ -18,6 +18,6 @@ ng_project(
         "//packages/core",
         "//packages/core/testing",
         "//packages/platform-browser",
-        "//packages/platform-server:bundled_domino_lib",
+        "//packages/platform-server/third_party/domino:bundled_domino_lib",
     ],
 )
