+++ tools/testing/node_zoneless_tests.init.mts
@@ -16,7 +16,7 @@ import {TestBed} from '@angular/core/testing';
 import {NgModule, provideZonelessChangeDetection} from '@angular/core';
 import {ServerTestingModule, platformServerTesting} from '@angular/platform-server/testing';
 import {ɵDominoAdapter as DominoAdapter} from '@angular/platform-server';
-import domino from '../../packages/platform-server/src/bundled-domino';
+import domino from '../../packages/platform-server/third_party/domino/bundled-domino';
 
 @NgModule({
   providers: [provideZonelessChangeDetection()],
